Add func for creating a new user with curl in console
