+++ jwt/types.go
@@ -5,6 +5,7 @@ import (
 )
 
 type Session struct {
+	Login   bool   `json:"login,omitempty"`
 	ID      int64  `json:"id,omitempty"`
 	Expired int64  `json:"exp,omitempty"`
 	GroupID uint32 `json:"grp,omitempty"`
