+++ spa/src/_api/payments.js
@@ -27,9 +27,9 @@ export default class {
     return Server.get(`drivers/${id}`);
   }
 
-  static getUserByLogin (login){
-    return Server.get(`driver/${login}`);
-  }
+  // static getUserByLogin (login){
+  //   return Server.get(`driver/${login}`);
+  // }
 
   static auth(login, password) {
     return Server.post(`login`, {
