+++ spa/src/containers/Payments/index.js
@@ -20,7 +20,7 @@ class Index extends Component {
   };
 
   render() {
-    console.log('in paymets,this.props.isAuth',this.props.isAuth);
+    // console.log('in paymets,this.props.isAuth',this.props.isAuth);
     return <div>
       <div style={{padding: '10px 5px'}}>
         <StatementSelect
@@ -47,7 +47,7 @@ const mapStateToProps = state => ({
   payments: state.payments.list,
   statements: state.payments.statements,
   statementUUID: state.payments.statementUUID,
-  isAuth: state.global.isAuth,
+  // isAuth: state.global.isAuth,
 
 });
